+++ rxjava-core/src/test/java/rx/ConcatTests.java
@@ -7,6 +7,12 @@ import java.util.List;
 
 import org.junit.Test;
 
+import rx.CovarianceTest.HorrorMovie;
+import rx.CovarianceTest.Media;
+import rx.CovarianceTest.Movie;
+import rx.Observable.OnSubscribeFunc;
+import rx.subscriptions.Subscriptions;
+
 public class ConcatTests {
 
     @Test
@@ -53,5 +59,63 @@ public class ConcatTests {
         assertEquals("two", values.get(1));
         assertEquals("three", values.get(2));
         assertEquals("four", values.get(3));
+    }
+    
+    @Test
+    public void testConcatCovariance() {
+        Observable<Media> o1 = Observable.<Media> from(new HorrorMovie(), new Movie());
+        Observable<Media> o2 = Observable.from(new Media(), new HorrorMovie());
+
+        Observable<Observable<Media>> os = Observable.from(o1, o2);
+
+        List<Media> values = Observable.concat(os).toList().toBlockingObservable().single();
+    }
+
+    @Test
+    public void testConcatCovariance2() {
+        Observable<Media> o1 = Observable.from(new HorrorMovie(), new Movie(), new Media());
+        Observable<Media> o2 = Observable.from(new Media(), new HorrorMovie());
+
+        Observable<Observable<Media>> os = Observable.from(o1, o2);
+
+        List<Media> values = Observable.concat(os).toList().toBlockingObservable().single();
+    }
+    
+    @Test
+    public void testConcatCovariance3() {
+        Observable<Movie> o1 = Observable.from(new HorrorMovie(), new Movie());
+        Observable<Media> o2 = Observable.from(new Media(), new HorrorMovie());
+
+        List<Media> values = Observable.concat(o1, o2).toList().toBlockingObservable().single();
+        
+        assertTrue(values.get(0) instanceof HorrorMovie);
+        assertTrue(values.get(1) instanceof Movie);
+        assertTrue(values.get(2) instanceof Media);
+        assertTrue(values.get(3) instanceof HorrorMovie);
+    }
+
+    @Test
+    public void testConcatCovariance4() {
+
+        Observable<Movie> o1 = Observable.create(new OnSubscribeFunc<Movie>() {
+
+            @Override
+            public Subscription onSubscribe(Observer<? super Movie> o) {
+                o.onNext(new HorrorMovie());
+                o.onNext(new Movie());
+                //                o.onNext(new Media()); // correctly doesn't compile
+                o.onCompleted();
+                return Subscriptions.empty();
+            }
+        });
+
+        Observable<Media> o2 = Observable.from(new Media(), new HorrorMovie());
+
+        List<Media> values = Observable.concat(o1, o2).toList().toBlockingObservable().single();
+        
+        assertTrue(values.get(0) instanceof HorrorMovie);
+        assertTrue(values.get(1) instanceof Movie);
+        assertTrue(values.get(2) instanceof Media);
+        assertTrue(values.get(3) instanceof HorrorMovie);
     }
 }
